Calculate Timer image fill value
Refactored code to be the same as the tutorial. Will change it back
later
Calculate a fillFraction value between 0 and 1 which will be used to
show the remaining time in our timer image
Created a variable to inform us to load the next question
Created a method to cancel the timer early for when the user answered
the question.