Added End Screen
Added End Screen to game showing the final score of the player and asks whether the
player wants to play again.
Added a GameManager which controls the flow of the canvases, i.e. which
screen is being shown at what time.
Game tutorial is now complete but there are some things to come back to
later.